a4d/requirements.txt
django==4.*
django-rest-framework
Pillow
